+++ src/phantasyfootballer/run.py
@@ -1,38 +1,12 @@
-# Copyright 2020 QuantumBlack Visual Analytics Limited
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
-# EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
-# OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND
-# NONINFRINGEMENT. IN NO EVENT WILL THE LICENSOR OR OTHER CONTRIBUTORS
-# BE LIABLE FOR ANY CLAIM, DAMAGES, OR OTHER LIABILITY, WHETHER IN AN
-# ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF, OR IN
-# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-#
-# The QuantumBlack Visual Analytics Limited ("QuantumBlack") name and logo
-# (either separately or in combination, "QuantumBlack Trademarks") are
-# trademarks of QuantumBlack. The License does not grant you any right or
-# license to the QuantumBlack Trademarks. You may not use the QuantumBlack
-# Trademarks or any confusingly similar mark as a trademark for your product,
-# or use the QuantumBlack Trademarks in any other manner that might cause
-# confusion in the marketplace, including but not limited to in advertising,
-# on websites, or on software.
-#
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 """Application entry point."""
 from pathlib import Path
 from typing import Dict
-
+import os
+from typing import Optional
 from kedro.framework.context import KedroContext, load_package_context
 from kedro.pipeline import Pipeline
 from phantasyfootballer.pipeline import create_pipelines
+from phantasyfootballer.pipelines.full_pipeline import create_full_pipeline
 
 
 class ProjectContext(KedroContext):
@@ -45,8 +19,28 @@ class ProjectContext(KedroContext):
     project_version = "0.16.4"
     package_name = "phantasyfootballer"
 
+    def __init_(
+        self, project_path: str, env: Optional[str] = os.getenv("PYTHON_ENV"), **kwargs
+    ):
+        super().__init__(project_path, env, **kwargs)
+        ## Any other properties that need to be set for this project context, which will be shipped around - especially helpful in analysis
+
+    @property
+    def pipeline(self):
+        """ Default pipeline for the PhantasyFootballer app
+        """
+        return create_full_pipeline()
+
     def _get_pipelines(self) -> Dict[str, Pipeline]:
+        # TODO: Here I can override how the pipelines are created, might be helpful to set boundaries on the data used
         return create_pipelines()
+
+    def _get_catalog(self, save_version=None, journal=None, load_versions=None):
+        # TODO: This is the spot where I can add data to the catalog when instantiated
+        catalog = super()._get_catalog(
+            save_version=save_version, journal=journal, load_versions=load_versions
+        )
+        return catalog
 
 
 def run_package():