+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [0.3.6] - 2020-09-06
 
+### [Unreleased]
+- Modified the file structure, and data catalog identifiers for annual projections
+
+-[Breaking] removed the possibility of there being a "default" pipeline
+
 ### Changed
 
-- TODO
 
 ## [Version] - 2020-09-06